+++ src/jetty-chart/src/common/line-common/utils/calulate-curve.js
